requirements.txt
a2a==0.44
a2a-sdk==0.3.22
ag-ui-protocol==0.1.10
agentscope==1.0.11
agentscope-runtime==1.0.5.post1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
alibabacloud-credentials==1.0.7
alibabacloud-credentials-api==1.0.0
alibabacloud-darabonba-string==0.0.4
alibabacloud-gateway-pop==0.1.3
alibabacloud-kms20160120==2.4.0
alibabacloud-openapi-util==0.2.4
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.3
alibabacloud-tea-util==0.3.14
alibabacloud_darabonba_array==0.1.0
alibabacloud_darabonba_encode_util==0.0.2
alibabacloud_darabonba_map==0.0.1
alibabacloud_darabonba_signature_util==0.0.4
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_tea_xml==0.0.3
# aliyun-python-sdk-core==2.16.0
# aliyun-python-sdk-core>=2.13.0
# aliyun-python-sdk-kms==2.16.5
# aliyun-python-sdk-core>=2.13.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.1
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
asttokens
async-timeout==5.0.1
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
bidict==0.23.1
billiard==4.2.4
brotli==1.2.0
cached-property==2.0.1
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm
constantly==23.10.4
crcmod==1.7
cryptography==44.0.3
cssselect==1.4.0
darabonba-core==1.0.5
dashscope==1.25.11
ddgs==9.10.0
debugpy
decorator
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
exceptiongroup
executing
fake-useragent==2.2.0
fastapi==0.128.1
filelock==3.20.3
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.48.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
ipykernel
ipython
itemadapter==0.13.1
itemloaders==1.4.0
jedi
Jinja2==3.1.6
jiter==0.13.0
# jmespath>=0.9.3,<2.0.0
# aliyun-python-sdk-core>=2.13.0
# aliyun-python-sdk-core>=2.13.0 --no-deps
json5==0.13.0
json_repair==0.56.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client
jupyter_core
kombu==5.6.2
kubernetes==35.0.0
langchain==1.2.8
langchain-core==1.2.8
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
# lxml==6.0.2
lxml>=4.9.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
nacos-sdk-python==3.0.4
nest_asyncio
numpy==2.2.6
oauthlib==3.3.1
openai==2.16.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
# oss2==2.19.1
packaging
parsel==1.11.0
parso
pexpect
pickleshare
platformdirs
primp==0.15.0
prompt_toolkit
propcache==0.4.1
Protego==0.6.0
proto-plus==1.27.1
protobuf==6.33.5
psutil
ptyprocess
pure_eval
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyDispatcher==2.0.7
Pygments
PyJWT==2.11.0
pyOpenSSL==25.1.0
python-datauri==3.0.2
python-dateutil
python-dotenv==1.2.1
python-engineio==4.13.0
python-frontmatter==1.1.0
python-multipart==0.0.22
python-socketio==5.16.0
PyYAML==6.0.3
pyzmq
queuelib==1.9.0
redis==6.4.0
referencing==0.37.0
regex==2026.1.15
readability-lxml==0.8.4.1
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
Scrapy==2.14.1
service-identity==24.2.0
shortuuid==1.0.13
simple-websocket==1.1.0
six
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.5
soupsieve==2.8.3
sse-starlette==3.2.0
stack_data
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tldextract==5.3.1
tomli==2.4.0
tornado
tqdm==4.67.3
traitlets
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
# uvloop==0.22.1 ; sys_platform != "win32"
vine==5.1.0
w3lib==2.4.0
watchfiles==1.1.1
wcwidth
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface==8.2
zstandard==0.25.0
